+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+### Added
+
+* `ImmersClient.logout` gains optional `alsoLogoutFromImmer` argument to fully terminate login sessions
+for users from the local immer
+
 ## v2.15.0 (2023-03-10)
 
 ### Added